Encode rlfx using hierarchical automata syntax
See comment in `pin_automaton.lus` for link to syntax. The encoding
appears to work, but I don't currently know how to talk about current
and next state in properties, so several cannot be posed at the moment.
Also, the analysis appears to be much slower - the underlying encoding
from the lus into kind2 seems to be much more complex.